+++ trixnity-messenger-compose-view/src/commonMain/kotlin/de/connect2x/trixnity/messenger/compose/view/i18n/I18n.kt
@@ -3246,7 +3246,7 @@ open class I18nView(
 
     open fun knockLabel() = translate {
         EN - "Why do you wish to join? (Optional)"
-        DE - "Warum möchtest du beitreten? (Optional)"
+        DE - "Warum möchten Sie beitreten? (Optional)"
     }
 
     open fun allRoomMembers() = translate {
@@ -3715,28 +3715,37 @@ open class I18nView(
     }
 
     open fun joinRoomActionKnock() = translate {
-        EN - "Without an invitation you need to knock on the room and be let in. Please confirm you want to knock."
+        EN -
+            "Without an invitation, you need to request to join the room and be let in. Please confirm you want to send a join request."
         DE -
-            "Um ohne eine Einladung beizutreten, müssen Sie eine Beitrittsanfrage stellen und hereingelassen werden. Bitte bestätigen Sie das Anklopfen."
+            "Um ohne Einladung beizutreten, müssen Sie eine Beitrittsanfrage stellen und eingelassen werden. Bitte bestätigen Sie, dass Sie eine Beitrittsanfrage stellen möchten."
     }
 
     open fun joinRoomActionKnockSuccess() = translate {
         EN -
-            "Successfully knocked, you will be redirected to the room once the request is accepted. Alternatively you can check back later."
+            "Join request sent. You will be redirected to the room once it is accepted. Alternatively, you can check back later."
         DE -
-            "Die Beitrittsanfrage war erfolgreich. Der Raum wird geöffnet, sobald die Anfrage akzeptiert wurde. Andernfalls können Sie das Öffnen später erneut versuchen. "
+            "Die Beitrittsanfrage wurde gesendet. Sie werden zum Raum weitergeleitet, sobald sie angenommen wurde. Alternativ können Sie später erneut nachsehen."
     }
 
     open fun joinRoomActionRestricted(requiredRooms: Set<RoomAliasId>) = translate {
         EN -
-            "Without an invitation you need to be a member of at least one of the following rooms to join: ${requiredRooms.joinToString()}."
+            "Join this room to access its content. Without an invitation, you must be a member of at least one of these rooms: ${requiredRooms.joinToString()} to do so."
         DE -
-            "Um ohne eine Einladung beizutreten, müssen Sie Mitglied in mindestens einem der folgenden Räume sei: ${requiredRooms.joinToString()}."
+            "Treten Sie diesem Raum bei, um seinen Inhalt zu sehen. Ohne Einladung müssen Sie hierzu Mitglied in mindestens einem dieser Räume sein: ${requiredRooms.joinToString()}."
     }
 
-    open fun joinRoomActionRestrictedNoRoomInfoAvailable() = translate {
-        EN - "Without an invitation you need to be a member of another room."
-        DE - "Um ohne eine Einladung beizutreten, müssen Sie bereits Mitglied eines anderen Raumes sein."
+    open fun joinRoomActionRestricted(requiredUnknownRooms: Int) = translate {
+        EN - "To access this room's content, get invited or be a member of at least one other unknown room."
+        DE -
+            "Um auf den Inhalt dieses Raums zuzugreifen, lass dich einladen oder sei Mitglied in mindestens einem weiteren unbekannten Raum."
+    }
+
+    open fun joinRoomActionRestricted(requiredRooms: Set<RoomAliasId>, requiredUnknownRooms: Int) = translate {
+        EN -
+            "To access this room's content, get invited, be a member of any of ${requiredRooms.joinToString()}, or be a member of at least one other unknown room."
+        DE -
+            "Um auf den Inhalt dieses Raums zuzugreifen, lass dich einladen, sei Mitglied in einem der Räume ${requiredRooms.joinToString()} oder in mindestens einem weiteren unbekannten Raum."
     }
 
     open fun joinRoomActionAcceptInvite() = translate {
@@ -3751,6 +3760,51 @@ open class I18nView(
             "Couldn't find the mentioned room. Either the id isn't valid or the homeserver doesn't allow checking the join rules of rooms. Please request an invitation to the room instead."
         DE -
             "Der Raum konnte nicht gefunden werden. Das liegt daran, dass die id des Raumes ungültig ist oder die Beitrittsregeln des Raumes nicht öffentlich sind. Bitte fordern Sie stattdessen eine Einladung in den Raum an."
+    }
+
+    open fun rejoinRoomJoinButton() = translate {
+        EN - "Rejoin room"
+        DE - "Raum erneut beitreten"
+    }
+
+    open fun rejoinRoomJoin() = translate {
+        EN - "Rejoin this room to see newer messages."
+        DE - "Treten Sie diesem Raum erneut bei, um neuere Nachrichten zu sehen."
+    }
+
+    open fun rejoinRoomKnockButton() = translate {
+        EN - "Request rejoin"
+        DE - "Wiederbeitritt anfragen"
+    }
+
+    open fun rejoinRoomKnock() = translate {
+        EN - "Rejoin this room to see newer messages."
+        DE - "Treten Sie diesem Raum erneut bei, um neuere Nachrichten zu sehen."
+    }
+
+    open fun rejoinRoomKnockKnocked() = translate {
+        EN - "Join request sent. Once it is accepted, you will be a member of this room again."
+        DE - "Die Beitrittsanfrage wurde gesendet. Sobald sie angenommen wurde, sind sie wieder Mitglied dieses Raumes."
+    }
+
+    open fun rejoinRoomRestricted(requiredRooms: Set<RoomAliasId>) = translate {
+        EN -
+            "To read newer messages, get invited again or be a member of at least one of these rooms: ${requiredRooms.joinToString()}."
+        DE -
+            "Um neuere Nachrichten zu lesen, lass dich erneut einladen oder sei Mitglied in mindestens einem dieser Räume: ${requiredRooms.joinToString()}."
+    }
+
+    open fun rejoinRoomRestricted(requiredUnknownRooms: Int) = translate {
+        EN - "To read newer messages get invited again or be a member of at least one other unknown room."
+        DE -
+            "Um neuere Nachrichten zu lesen, lass dich erneut einladen oder sei Mitglied in mindestens einem weiteren unbekannten Raum."
+    }
+
+    open fun rejoinRoomRestricted(requiredRooms: Set<RoomAliasId>, requiredUnknownRooms: Int) = translate {
+        EN -
+            "To read newer messages get invited again, be a member of any of ${requiredRooms.joinToString()}, or be a member of at least one other unknown room."
+        DE -
+            "Um neuere Nachrichten zu lesen, lass dich erneut einladen, sei Mitglied in einem der Räume ${requiredRooms.joinToString()} oder in mindestens einem weiteren unbekannten Raum."
     }
 
     open fun userSearchFilter() = translate {